Finished Test cleanup. Now only missing 4 test for logo.

+++ src/eDoxa.Organizations.Clans.Api/Areas/Clans/Services/CandidatureService.cs
@@ -1,6 +1,6 @@
 ﻿// Filename: CandidatureService.cs
 // Date Created: 2019-09-30
-// 
+//
 // ================================================
 // Copyright © 2019, eDoxa. All rights reserved.
 
@@ -72,13 +72,16 @@ namespace eDoxa.Organizations.Clans.Api.Areas.Clans.Services
 
             candidature.Accept();
 
-            await _candidatureRepository.UnitOfWork.CommitAsync();
-
             return new ValidationResult();
         }
 
-        public async Task<ValidationResult> DeclineCandidatureAsync(Candidature candidature, UserId ownerId)
+        public async Task<ValidationResult> DeclineCandidatureAsync(Candidature candidature, UserId userId)
         {
+            if (!await _clanRepository.IsOwnerAsync(candidature.ClanId, userId))
+            {
+                return new ValidationFailure(string.Empty, "Permission required.").ToResult();
+            }
+
             _candidatureRepository.Delete(candidature);
 
             await _candidatureRepository.UnitOfWork.CommitAsync();
